+++ tests/integration/centralized.rs
@@ -52,6 +52,7 @@ mod other {
   centralized_test!(other, should_gracefully_quit);
   centralized_test!(other, should_support_options_with_pipeline);
   centralized_test!(other, should_reuse_pipeline);
+  centralized_test!(other, should_manually_connect_twice);
 
   //#[cfg(feature = "dns")]
   // centralized_test!(other, should_use_trust_dns);