+++ tests/integration/clustered.rs
@@ -56,6 +56,7 @@ mod other {
   cluster_test!(other, should_gracefully_quit);
   cluster_test!(other, should_support_options_with_pipeline);
   cluster_test!(other, should_reuse_pipeline);
+  cluster_test!(other, should_manually_connect_twice);
 
   //#[cfg(feature = "dns")]
   // cluster_test!(other, should_use_trust_dns);