+++ tests/integration/other/mod.rs
@@ -551,3 +551,26 @@ pub async fn should_reuse_pipeline(client: RedisClient, _: RedisConfig) -> Resul
   assert_eq!(pipeline.last::<i64>().await?, 4);
   Ok(())
 }
+
+pub async fn should_manually_connect_twice(client: RedisClient, _: RedisConfig) -> Result<(), RedisError> {
+  let client = client.clone_new();
+  let _old_connection = client.connect();
+  let _ = client.wait_for_connect().await?;
+
+  let blpop_jh = tokio::spawn({
+    let client = client.clone();
+    async move { client.blpop::<Option<i64>, _>("foo", 30.0).await }
+  });
+
+  // TODO check what happens if you abort the old connection before calling connect again
+
+  sleep(Duration::from_millis(100)).await;
+  let new_connection = client.connect();
+  let _ = client.wait_for_connect().await?;
+  assert_eq!(blpop_jh.await.unwrap(), Err(RedisError::new_canceled()));
+
+  assert_eq!(client.incr::<i64, _>("bar").await?, 1);
+  let _ = client.quit().await?;
+  let _ = new_connection.await?;
+  Ok(())
+}